# specs/006-chatkit-frontend/contracts/api.yaml
openapi: 3.0.3
info:
  title: Todo AI Chatbot - Task List API
  description: API endpoint for fetching user tasks for sidebar dashboard
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Development server

paths:
  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      description: Returns a list of all tasks belonging to the specified user, ordered by creation date descending.
      operationId: getTasks
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User's unique identifier
          schema:
            type: string
            minLength: 1
            example: "demo-user"
      responses:
        '200':
          description: Successfully retrieved task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
              example:
                tasks:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    user_id: "demo-user"
                    title: "Buy groceries"
                    description: "Milk, eggs, bread"
                    completed: false
                    due_date: "2026-01-25T17:00:00Z"
                    priority: "high"
                    category: "Shopping"
                    created_at: "2026-01-22T10:00:00Z"
                    updated_at: "2026-01-22T10:00:00Z"
                  - id: "550e8400-e29b-41d4-a716-446655440001"
                    user_id: "demo-user"
                    title: "Review PR"
                    description: null
                    completed: true
                    due_date: null
                    priority: "medium"
                    category: "Work"
                    created_at: "2026-01-21T09:00:00Z"
                    updated_at: "2026-01-22T08:30:00Z"
        '400':
          description: Bad request - invalid user_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "user_id is required and cannot be empty"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Failed to fetch tasks"

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        user_id:
          type: string
          description: Owner's user identifier
          example: "demo-user"
        title:
          type: string
          maxLength: 255
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          description: Detailed task description
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Completion status
          example: false
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Task due date in ISO8601 format
          example: "2026-01-25T17:00:00Z"
        priority:
          type: string
          nullable: true
          enum:
            - high
            - medium
            - low
          description: Task priority level
          example: "high"
        category:
          type: string
          nullable: true
          maxLength: 50
          description: User-defined category
          example: "Shopping"
        created_at:
          type: string
          format: date-time
          description: Creation timestamp in ISO8601 format
          example: "2026-01-22T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last modification timestamp in ISO8601 format
          example: "2026-01-22T10:00:00Z"

    TaskListResponse:
      type: object
      required:
        - tasks
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
          description: List of tasks for the user

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "An error occurred"